requirements.txt
asgiref==3.8.1
coverage==7.5.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
